+++ Assets/2DGame/Scripts/DataPersistentSystem/DataPersistentUtility.cs
@@ -14,14 +14,13 @@ public static class DataPersistentUtility
     public static string ExtractAssetName(ScriptableObject item)
     {
         string itemName = item.ToString();
-        int startIndex = 0; // trouver la position du premier underscore
-        int endIndex = itemName.IndexOf(" (", StringComparison.Ordinal) - 1; // trouver la position de la parenthèse ouvrante
-        int length = endIndex  + 1; // calculer la longueur de la sous-chaîne
-        string finalName = itemName.Substring(startIndex, length); // extraire la sous-chaîne
+        int startIndex = 0; 
+        int endIndex = itemName.IndexOf(" (", StringComparison.Ordinal) - 1; 
+        int length = endIndex  + 1;
+        string finalName = itemName.Substring(startIndex, length);
          
         return finalName;
     }
-
     public static T GetAssetFromResources<T>(string assetName) where T : Object
     {
         if (string.IsNullOrEmpty(assetName))
@@ -33,10 +32,8 @@ public static class DataPersistentUtility
         {
             Debug.LogError($"Asset {assetName} is not found, you need to create the asset or to add it in a Resources Folder");
         }
-
         return asset;
     }
-
     public static int GenerateID()
     {
         return Guid.NewGuid().GetHashCode();
