+++ Assets/2DGame/Scripts/DataPersistentSystem/ISaveInstance.cs
@@ -1,0 +1,13 @@
+/// <summary>
+///  <inheritdoc/>
+/// ISaveInstance exist for instancied component on a scene
+/// </summary>
+public interface ISaveInstance : ISave
+{
+    /// <summary>
+    /// Unique ID, required because it distinct different saved instance from the same MonoBehavior
+    /// </summary>
+    public int SaveID { get; }
+}
+
+
