+++ Assets/2DGame/Scripts/DataPersistentSystem/SaveData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+/// <summary>
+/// Abstract class used to be converted in JsonString, you need to create a child class to add complementary property to save
+/// </summary>
+[Serializable]
+public abstract class SaveData
+{
+    // TODO : Add scene management ? IDK we need to see
+}
