+++ Assets/2DGame/Scripts/DataPersistentSystem/ScriptableObjectSaveable.cs
@@ -2,7 +2,10 @@
 
 namespace _2DGame.Scripts.Save
 {
-    public abstract class ScriptableObjectSaveable : ScriptableObject, ISaveData
+    /// <summary>
+    /// ScriptableObject compatible with DataPersistantSystem
+    /// </summary>
+    public abstract class ScriptableObjectSaveable : ScriptableObject, ISave
     {
         public abstract void OnLoad(string data);
         public abstract void OnSave(out SaveData saveData);
