--- Assets/2DGame/Scripts/Save/ISaveData.cs
@@ -1,7 +1,0 @@
-public interface ISaveData
-{
-    public void OnLoad(string data);
-    public void OnSave(out SaveData saveData);
-}
-
-
