--- Assets/2DGame/Scripts/Save/ISaveInstance.cs
@@ -1,7 +1,0 @@
-public interface ISaveInstance : ISaveData
-{
- 
-    public int SaveID { get; }
-}
-
-
