--- Assets/2DGame/Scripts/Save/SaveData.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-[Serializable]
-public class SaveData
-{
-    //public string type;
-
-    public SaveData()
-    {
-       // type = nameof(SaveData);
-    }
-}
